refactor: replace old implementation of devel-project.py
